backend/requirements.txt
django==3.2.5
djangorestframework==3.12.4
pytest-django==4.4.0
pytest==6.2.4
djoser==2.1.0
Pillow==8.3.1
django-filter==2.4.0